refactor(core): extract shared StorageBase from RolloutStore
First step of #214. Lifts the schema-agnostic half of `RolloutStore` into a
new `store_base` module so `ContextStore` and `DatagenStore` can be migrated
onto it in follow-ups, instead of each carrying its own forked copy.

`StorageBase` owns everything that is not schema-specific:

- open/create, and the `load_with_options` / `create_with_options` /
  `DatasetNotFound`-fallback triad that is currently written out four times
  across the tree
- the resident `ShardWriter` lifecycle: lazy open, fence detection and
  single retry, identity-checked invalidation, `close()`, and the `Drop`
  fallback that seals an evicted store's memtable
- `put` (durable append, no seal) and `flush` (seal + drain), with the
  add/flush latency and error metrics
- the WAL merge: prepare/commit split, surgical generation drain reusing the
  shard's current epoch, post-drain directory deletion, and both the count
  and time triggers
- compaction with `defer_index_remap`, ZoneMap index management, MemWAL
  index initialization, shard-manifest discovery, and the LSM scanner

`RolloutStore` is now the rollout *schema* plus encode/decode and the typed
read APIs, delegating storage to its `base` field. Behavior is unchanged --
this commit is a pure extraction, and rollout is the reference implementation
the other two stores will be moved onto.

Also deduplicates `align_batch_to_schema`, `derive_shard_id`,
`is_fenced_error`, `is_not_found_error`, `ListSource` and `PreparedMerge`
into `store_base`; `rollout_store` re-exports them so the public API is
unchanged. `DatagenStore` now imports them from `store_base` rather than
reaching into `rollout_store`.

Net -1145 lines. Full workspace test suite passes unchanged (178 core lib
tests, WAL merge concurrency and generation-cleanup integration tests),
clippy is clean, and rustdoc warnings drop from 13 to 7.
